[timer] support negative aDt for StartAt.
This PR adds support for negative value `aDt`. This is useful when some
margin ahead is wanted when schedule a timer.

`kForeverDt` is adjusted to `(1UL<<31)` instead of `0xffffffff`(or `-1`)
which is valid negative value for `aDt`.

Unit test is also enhanced to cover negative `aDt`.

+++ src/core/mac/sub_mac.cpp
@@ -447,7 +447,7 @@ void SubMac::SampleRssi(void)
 
     if (TimerMilliScheduler::IsStrictlyBefore(TimerMilli::GetNow(), mEnergyScanEndTime))
     {
-        mTimer.StartAt(mTimer.GetFireTime(), kEnergyScanRssiSampleInterval);
+        mTimer.StartAt(mTimer.GetFireTime(), static_cast<int32_t>(kEnergyScanRssiSampleInterval));
     }
     else
     {
